Implement endpoint for creating an Idea
This commit focuses on creating a post endpoint that creates an idea from
the parameters supplied via the request.

It is important to note that this commit is the first to follow the initial
setup commit.

Withing we have included and used the following gems:

- active_model_serializers _
- rspec (for testing)
- simplecov (tracking test coverage)
- json_matchers (match json responses to pre defined schema)

Routes defined is POST /v1/ideas which points to the ideas controller create action..