Implement validation for User entity
This updates the code to include and cater for scenarios where attributes
do not conform to the expected form or value.

It worth noting that the validation messages have been modified via the locales
YAML file as defined in [1]. [1] Provides guidance on the right property name
that overrides the defaults as defined in the rails gem.

[1] https://github.com/rails/rails/blob/master/activemodel/lib/active_model/locale/en.yml